+++ packages/catalyst/src/commands/build.spec.ts
