+++ packages/catalyst/src/commands/version.spec.ts
