+++ packages/catalyst/src/lib/project-config.spec.ts
