+++ packages/catalyst/vitest.config.ts
@@ -3,12 +3,12 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     environment: 'node',
-    setupFiles: './tests/vitest.setup.ts',
-    include: ['tests/**/*.spec.ts'],
+    setupFiles: './vitest.setup.ts',
+    include: ['src/**/*.spec.ts'],
     exclude: ['**/node_modules/**', '**/dist/**'],
     coverage: {
       include: ['src/**/*.ts'],
-      exclude: ['src/index.ts'],
+      exclude: ['src/cli/index.ts'],
       thresholds: {
         100: true,
       },
